+++ odoo/local-src/connector_qoqa/product_attribute_value/common.py
@@ -2,7 +2,7 @@
 # © 2013-2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
-from openerp import models, fields
+from openerp import _, api, models, exceptions, fields
 
 from ..unit.backend_adapter import QoQaAdapter
 from ..backend import qoqa
@@ -30,6 +30,21 @@ class ProductAttributeValue(models.Model):
         string='Attribute Bindings',
         copy=False,
     )
+
+    @api.constrains('name')
+    def _check_name_length(self):
+        for record in self:
+            if not record.name:
+                continue
+            # This is a business request. We don't limit the 'size' on the
+            # field because historic attribute values might be longer.
+            # When an historic value is longer, we expect to have an error
+            # so the user has to correct it.
+            if len(record.name) > 25:
+                raise exceptions.ValidationError(
+                    _("The limit for an attribute value is 25 characters.\n"
+                      "'%s' should therefore be shorter.") % (record.name,)
+                )
 
 
 @qoqa